refactor(tape): align service ownership
